refactored and now emails are not to be repeated

+++ src/containers/actions/auth.js
@@ -38,7 +38,6 @@ export const auth = (email, password, isSignup, token, userId) => {
     const getAuth = authData;
        axios.post(url, authData)
             .then(response => {
-                dispatch(authSuccess(response.data));
                   if( getAuth === authData){  
                     dispatch(authSuccess(response.data));
                     axios.get(url)
@@ -70,40 +69,78 @@ export const auth = (email, password, isSignup, token, userId) => {
 
                  })
                 //console.log(formMatch);
-                     const number = Object.keys(formMatch)
 
+
+                const wrapper = (counterIndex) => {
+                    // const number = Object.keys(formMatch)
+                     counterIndex = 0;
 
                     //console.log("form map: " , [...formMatch]
                         const copyDelete = [];
                         //const copy = [];
-                       for(let i = 0;  i < formMatch.length; i++){
+                       for(let a = 0;  a < formMatch.length; a++){
                      //  console.log("this is number:", parseInt(number[i]))
-                            copyDelete.push(formMatch[i]);
+                            copyDelete.push(formMatch[a]);
                           } 
-                            formMatch.forEach((item, i) => {
-                                let length = Number(number[i]);
-                                //console.log(length)
-                                //let newI = i + 1;
-                                let j = length - 1;
-                                let a = length - 2;
-                             // console.log(newI)
-                             if(i > 0){
-                                if(item === copyDelete[j] || item === copyDelete[0] || item === copyDelete[a]){
-                                   //console.log("this", copyDelete[length-1])
-                                   // console.log("this", item)
-                                    alert("stop! you already send this email.")
-                                    let userArray = `https://aimee-github.firebaseio.com/sign-up/${formMatchId[i]}.json`;
-                                    axios.delete(userArray); 
-                                    console.log(userArray);
+
+                            /* formMatch.forEach((item, i) => { */
+                                for(let i = 0;  i < formMatch.length; i++){
+                                counterIndex += 1;
+                               // console.log("wrapper: ", counterIndex)
+                                let maxLength = Number(counterIndex);
+                                console.log("This is number of indexes signed:", maxLength);
+                               // let length = Number(number[i]);
+                                //let lengthMax = Number(number[i]);
+                               // console.log("get the length:", formMatch.length);
+                                //console.log("this is i:", i);  
+                                //length -= 1;
+                              //console.log("this is length", length)
+                                //lengthMax -= 2;
+                               // console.log("this is 2 array find:", lengthMax);
+                                //let j = length - 1;
+                                //let a = length - 2;
+                              //console.log("this is j", j)
+                              //console.log("this is a", a)
+                               // console.log("this is copy delete:",copyDelete[i-1])
+                                //console.log("this is copy delete:", formMatch[i])
+                            const state = [];
+                            console.log("# data:")
+                               //console.log("this copy delete", copyDelete)
+                                for (let obj in copyDelete) {
+                                   // let i = Object.keys(copyDelete).indexOf(obj);    
+                                   console.log(obj)
+                                    state.push(obj)
                                 }
+                               // console.log("this length", copyDelete[i-1])
+                               //console.log("this form match length", formMatch[formMatch.length-1])
+             
+                                    let formBool =  formMatch[formMatch.length-1] !== copyDelete[i-1];
+                                   // let formBoolBack =  formMatch[i-2] !== copyDelete[i];
+                                    //let formBoolBackMore =  formMatch[i-3] !== copyDelete[i];
+                                    //let formBoolBackMoreFirst =  formMatch[i-4] !== copyDelete[i];
+
+                                   const getOnce = () => {
+                                   // let counter = 0;
+                                   // let callCounter = counter += 1;
+                                   // console.log(formBool)
+                                    if(!formBool  && i > 0){
+                                        alert("stop! you already send this email.")
+                                        let userArray = `https://aimee-github.firebaseio.com/sign-up/${formMatchId[i-1]}.json`;
+                                        axios.delete(userArray); 
+                                       // console.log(userArray);
+                                       //console.log("this is an error counter:" , callCounter)
+                                    }
+                                   }
+                                   getOnce()
+                                 
                                   //console.log("this is item ", item)
                                   //copy.push(item);
                                   //console.log("this is copy",[copy]
-                                  if( i > 3)  {
+                               /*    if( i > 3)  {
                                       alert("sorry database is full")
                                       return null;
-                                  }                   
-                                }
+                                  }    */                
+  
                                 if( formMatch[i] === "" || formMatchPass[i] === "") {
                                     alert("this is empty!")
                                     let userArray = `https://aimee-github.firebaseio.com/sign-up/${formMatchId[i]}.json`;
@@ -111,18 +148,10 @@ export const auth = (email, password, isSignup, token, userId) => {
                                     console.log(userArray);
                                     return null;
                                 }
-  
-                        })
+                            }
                         
-                        
-
-                
-                    
-  
-
-                       
-   
-        
+                        //)
+                        //return dispatch(authSuccess(authData));
                              /*  let userArray = `https://aimee-github.firebaseio.com/sign-up/${formMatchId[i]}.json`;
                               axios.delete(userArray);
                               console.log("this is form array: ", userArray)
@@ -142,11 +171,14 @@ export const auth = (email, password, isSignup, token, userId) => {
                         
                     
     
-                 
+                             
 //console.log("this is true or false", form);
-               
+           }
+           wrapper()
                     })
                 }
+              
+                
             })
             .catch(err => {
                 console.log(err);
